Move things around. Move SS3 sample class to a new file
